fix goblin

+++ code/modules/mob/living/simple_animal/hostile/shitcur_goblin.dm
@@ -28,7 +28,8 @@
 
 /mob/living/simple_animal/hostile/shitcur_goblin/lose_target()
 	message_admins("Smiting shitcurity goblin was deleted due to a lack of valid target. Someone killed them first, or they ceased to exist.")
-	qdel(src) //so we dont shitcur after banning
+	if(!QDELETED(src))
+		qdel(src) //so we dont shitcur after banning
 
 /mob/living/simple_animal/hostile/shitcur_goblin/AttackingTarget()
 	stun_attack(target)
